create custom aappbar for book detaiks and create the book pic,name,auther name and rating
