--- lib/features/home/presention/views/widgets/customAppBar.dart
@@ -1,25 +1,0 @@
-import 'package:bookly_app/core/utils/assets.dart';
-import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-
-class CustomAppBar extends StatelessWidget {
-  const CustomAppBar({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final h = MediaQuery.of(context).size.height;
-    final w = MediaQuery.of(context).size.width;
-    return Padding(
-      padding: EdgeInsets.symmetric(horizontal: w * 0.05, vertical: h * 0.035),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Image.asset(AssetsData.logo, height: 20),
-          IconButton(
-              onPressed: () {},
-              icon: const Icon(FontAwesomeIcons.magnifyingGlass, size: 24)),
-        ],
-      ),
-    );
-  }
-}
